Add files via upload
EXECUTIVE SUMMARY 

This data analysis on Python explored different variables within the Medicare providers and organizations. The data came from North Carolina, Tennessee, and South Carolina. North Carolina has the most amount of organization Medicare spending focused largely on Clinical Laboratories and Ambulance Service Provider. On a provider level, most Medicare spending is used by the following top 5 provider types: Opthalmology, Internal Medicine, Hermatology- Oncology, Nurse Practitioner, and Family Practice. It’s nice to identify the major areas of need to further look into funding/cost cutting measures. I was personally surprised by how much Opthalmology and Internal Medicine was spending.
Unsurprisingly, the opiate with the highest total daily supply was Hydrocodone/Acetaminophen. This can be attributed to its easy access and low price. The top 10 most expensive drugs were also identified in the analysis which would be helpful to set prices for drugs covered by Medicare. Of course, more data on average income of the patients and their out-of-pocket costs is needed for this further policy analysis.
I also calculated mean cost of claims by city and state and South Carolina had the highest cost overall. Again, more information on average income of the residents and the size of the burden of this cost would be a good statistic to look at. It makes sense that there were drastically different rates of opiate prescriptions by city because cities have different demographics and hence different supply and demand factors. Knoxville, TN had the highest total day supply of Hydrocodone/Acetaminophen followed by Charlotte, NC. We can use this analysis to study the
city specific factors that are causing this. We should also compare this to national averages to see if this is a problem. It’s interesting that when I looked at the day supply per person, Knoxville was not at the top. Spartanburg, SC topped that statistic with an average of 12.66 units per day. This struck me the most because that is a horrifying statistic. People could easily overdose if they take that many pills in a day. Most prescriptions were opiates which makes me think there is a structural issue within our society. If people keep getting prescribed these addictive drugs, it’s hard to end the problem of opiate addiction. We should look at alternative medicine and non- opiate treatment options.